etl/requirements.txt
# Notebook ETL
ipython==7.22.0
nbconvert==6.0.*

# Sphinx docs rendering
myst-parser>=0.13.5
nbsphinx 
nbconvert
nbformat
numpydoc
pyyaml==5.4.1
requests
sphinx